Update Readme
So my first version was just a block of text. I thought to add in line breaks and change my article to a hyperlink - but only once I'd see what an ugly mess it was on the page. Somehow I don't think  my commit went through last time.
What I really want is to learn JavaScript and write so it fits that format.